🧬 Extract the logic about processing test cases from test data into a new single module to manage and maintain.
